encode() now returns error code if save_file() cannot open output file.
Before this change, encode() would return 0 whether or not the output
file was created.  There was no way for a caller to know when the
file could not be created (e.g. due to a non-existent parent directory).

Changed save_file() to return error 79 if it cannot open output file.
encode() checks return code from save_file() and passes any error value
back to the caller.