Guard some truncating casts with & 0xFF
The result of adding two unsigned chars is an int, so assigning this again
to unsigned char results in a narrowing cast. Afaik this is well-defined
by the standard but code compiled with cl with all runtime checks enabled
will trigger a breakpoint when e.g. scanline[i] + recon[i - bytewidth]
yields a number > 255. Fix this, and make it clear what's happening, by
and'ing with 0xFF (which should be discarded by any sane optimizer anyway).