Allow colortype enum values up to 255
The enum has names for valid PNG color types, but is intended to be able
to represent invalid byte values as well, which can be used to diagnose
corrupted PNG files. This make sit able again to represent any invalid byte by
adding a not-to-be-actually-used entry with numeric value 255, so usage in
range 0-255 is guaranteed by the compiler.

This makes it again possible to assign the byte value in lodepng_inspect
to it directly, verified with clang++ -fsanitize=undefined.

Also slightly update the order of assignments and error returns in
lodepng_inspect: assign all output values before error checking, do
error checking in the order the bytes appear in the PNG header.